Disable super.flushBuffer on response wrapper (#83)
### Comments
If `flushBuffer()` gets called on the ServletResponse, it commits changes to the response such that we can no longer change headers. However, wovnjava relies on changing `content-length` because we alter the size of the response.

Here is from [Oracle docs](https://docs.oracle.com/javaee/7/api/javax/servlet/ServletResponse.html#flushBuffer--):
```
void flushBuffer()
          throws IOException
Forces any content in the buffer to be written to the client. A call to this method
automatically commits the response, meaning the status code and headers will be written.
```

Using the library with a servlet deployed on WebLogic, the response was flushed before reaching our filter, which made us unable to change `content-length` to a bigger value. Therefore, the last part of the document was lost.