Document supportedLangs setting (#86)
`supportedLangs` is an important setting, and this adds instructions for it to the README.
In order to help configure it, we also include a list of languages that WOVN can support.

Also, `"en"` is no longer included in `supportedLangs` by default.

#### Future work
This setting should probably be a required setting. For now, we don't change configuration requirements.