Set content-type to UTF-8 also for fallback method (#90)
When translating by API, we set a meta tag to declare content-type UTF-8.

With this change, we insert this meta tag also when processing the request without using the API for translation. Any existing content-type declaration will be replaced. Whenever wovnjava intercepts a request with text-based response, the output encoding will be UTF-8.