[Refactor] Create PatternHandlers (#104)
This is the first of several PRs related to URL management refactoring that I will all merge into `refactor/url_management` before releasing.

I create three UrlLanguagePatternHandlers for doing URL manipulation

- PathUrlLanguagePatternHandler
- QueryUrlLanguagePatternHandler
- SubdomainUrlLanguagePatternHandler

The proper UrlLanguagePatternHandler is initialized at filter startup.

UrlLanguagePatternHandler should be responsible for

    getLang(url)
    removeLang(url)
    insertLang(url, lang)

In this PR, only the `getLang` operation is migrated to the UrlLanguagePatternHandlers. Its behavior from before remains unchanged. The other UrlLanguagePatternHandler functions are placeholders for now.

Many lines in test files are changed, but this is mostly because (1) initialization and Header constructor has changed, and (2) initialization may now throw ConfigurationError. The other change affecting tests is the removal of UrlPatternReg setting.

### Changes
* Create UrlLanguagePatternHandlers for managing URLs of urlPattern path, query, and subdomain
* Remove UrlPatternReg setting
* Hide regular expressions for URL management inside the UrlLanguagePatternHandlers
* Raise useful ConfigurationError at startup if library settings are invalid