Allow Lang lookup only by language code (#109)
Change to allow Lang lookup only by language code. This fixes some incorrect behaviors in URL handling. `Lang.java` is greatly simplified in the process.

Note that in the Lang initialization, information for the names of languages is kept in the file, but it is no longer being used by the library.

### Changes
* Lang object lookup only uses language code now
    - This affects URL matches and settings configuration
* Raise ConfigurationError if the setting for defaultLang or supportedLangs is invalid
    - This is added now because a configuration like `defaultLang=English` becomes invalid
* Access ISO639-1 language code for hreflang using a Lang field instead of conversion function
* (Whitespace changes) retab test files for UrlLanguagePatternHandlers