Process each request only once (#115)
WovnServletFilter may intercept a single request multiple times if the request is forwarded internally within the servlet container. For instance, this forwarding happens when WovnServletFilter forwards a request with language code to the same request without language code.

WovnServletFilter will always set HTTP header `X-Wovn-Handler` when intercepting a request. Therefore, with these changes we will check if that header is set on the response object before processing the request. If the header is set, we will ignore it and pass it through untouched.

This change is more important now than before because we now calculate the request language etc based on the client requested URL, not the current servlet URL.