[Fix] Send correct client request URL in translation API request (#119)
Update translation API request to use the computed client request URL to identify which page is being translated.

The API request puts this information in two places. It sends the full URL in the API body, and it sends just the path in the cache key. We therefore have to extract the path part, and a new helper function is added for this purpose.

We use quite a bit of regex manipulation for URLs at this point. A lot of it is unavoidable, of course, and I'm trying to think about how to keep it to a minimum. That said, as long as any regex manipulation is tested in isolation, it feels manageable.