Remove query setting (#122)
The purpose of the `query` setting is to identify unique pages when the specified query parameters are different. This is a setting that the customer configures for their Wovn project.

The standard behavior for our products is to include all query parameters in the request for page data, and the server will identify the page correctly. Therefore, wovnjava does not need to concern itself with the details of query parameter filtering, and we can remove the setting.

Existing configuration files do not need to be changed.

**Old behavior:** Send only the query parameters required to identify the page
**New behavior:** Send all query parameters always, and let the server pick query parameters for page identification