Do nothing for invalid or previously processed requests (#126)
For some received requests, the filter needs to do nothing.

If the request has already been processed by wovnjava (and redirected, such that we receive it a second time), we should do nothing. We can assume that all necessary URL manipulation etc is already taken care of.

Also, if the request URL does not match the set of URLs that wovnjava is configured to process, we should not attempt to manipulate the URL by removing language code. We should instead do nothing. Such requests may happen if `sitePrefixPath` is set, and it will be relevant for Custom Domain.

This PR also changes a couple of names to be more general, so that we can more easily apply them to Custom Domain.