Always forward to path in default language (#130)
Currently when we intercept a request, we sometimes forward it with a different path (internally in the servlet container). This happens when the request path is different from the path in default language, such as when receiving `site.com/en/page.html` and `urlPattern = path`. We will forward that request to `site.com/page.html`.

However when we _don't_ need to translate the incoming request, such as for an image resource, we currently never forward it. Instead, the wrapper for the request object will try to hide language codes when accessing information about the path.

Hiding language codes from the getters on the request object is difficult because of how the servlet context influences the correct response, particularly for getters on the URL path. This stackoverflow post illustrates the meaning of some of these getters https://stackoverflow.com/a/21046620.

### Solution
Whenever the request has a path that is different from the equivalent path in default language, we want to forward it to the default language. That way, the servlet container will create the correct request object, and we do not have to wrap the getters for path information.

We still want to wrap the information about hostname, so we keep the `@Override` functions for `getServerName` and `getRequestURL`.