Change removeLang to convertToDefaultLang (#134)
With Custom Domain, it does not always make sense to "remove lang code" to convert a URL to default language. Therefore, we change the name of `UrlLanguagePatternHandler.removeLang` to  `UrlLanguagePatternHandler.convertToDefaultLang`, and remove the language argument. One consequence of this is that the function can only affect URL language codes that match a supported language.