[Feature] Support langCodeAliases (#141)
`langCodeAliases` lets you specify the language identifier for your supported languages.

> For example, the default language identifier for English is `en`, such that a URL for your page in English language may look like `http://site.com/en/page`.
> With `langCodeAliases`, you can change the language identifier to `us`, for example. The resulting URL would look like `http://site.com/us/page` instead.

See README for more details.

### Comments
This feature is relevant for path, query, and subdomain url patterns. As a result, the methods `getLang`, `convertToDefaultLanguage`, and `convertToTargetLanguage` has changed quite a bit for these pattern handlers.

`langCodeAliases` is a two-in-one feature from an implementation perspective. Allowing alias for a target language is a simple change, we only need to replace the default language code for identifying the language.

When an alias is set for default language, it changes the rules for interpreting requests. Request URLs without an explicit language identifier will now be ignored.

`getLang` has been changed to return `null` _only for input URLs that cannot be intercepted_. For any URL that WovnServletFilter should process, `getLang` will either return default language or a target language. (The last pattern handler, `CustomDomainUrlLanguagePatternHandler` already conforms to this behavior.)

With the new behavior of `getLang` , there is no longer need for the method `canInterceptUrl`. It has been removed.